# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditVerification": {
        "additionalProperties": false,
        "properties": {
          "checked_records": {
            "title": "Checked Records",
            "type": "integer"
          },
          "first_invalid_record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Invalid Record Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "checked_records"
        ],
        "title": "AuditVerification",
        "type": "object"
      },
...
      "ComponentStatus": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status",
          "detail"
        ],
        "title": "ComponentStatus",
        "type": "object"
      },
...
      "MetricsSnapshot": {
        "additionalProperties": false,
        "properties": {
          "capacity_rejections_total": {
            "default": 0,
            "title": "Capacity Rejections Total",
            "type": "integer"
          },
          "idempotent_replays_total": {
            "default": 0,
            "title": "Idempotent Replays Total",
            "type": "integer"
          },
          "latency_p50_ms": {
            "default": 0,
            "title": "Latency P50 Ms",
            "type": "number"
          },
          "latency_p95_ms": {
            "default": 0,
            "title": "Latency P95 Ms",
            "type": "number"
          },
          "payload_rejections_total": {
            "default": 0,
            "title": "Payload Rejections Total",
            "type": "integer"
          },
          "rate_limited_total": {
            "default": 0,
            "title": "Rate Limited Total",
            "type": "integer"
          },
          "requests_by_method": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Requests By Method",
            "type": "object"
          },
          "requests_by_route": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Requests By Route",
            "type": "object"
          },
          "requests_in_flight": {
            "default": 0,
            "title": "Requests In Flight",
            "type": "integer"
          },
          "requests_total": {
            "default": 0,
            "title": "Requests Total",
            "type": "integer"
          },
          "responses_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Responses By Status",
            "type": "object"
          },
          "unhandled_errors_total": {
            "default": 0,
            "title": "Unhandled Errors Total",
            "type": "integer"
          }
        },
        "title": "MetricsSnapshot",
        "type": "object"
      },
...
      "OperationalStatus": {
        "additionalProperties": false,
        "properties": {
          "audit": {
            "$ref": "#/components/schemas/AuditVerification"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentStatus"
            },
            "title": "Components",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsSnapshot"
          },
          "phase4_agent_count": {
            "title": "Phase4 Agent Count",
            "type": "integer"
          },
          "phases": {
            "items": {
              "type": "integer"
            },
            "title": "Phases",
            "type": "array"
          },
          "shared_agent_count": {
            "title": "Shared Agent Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "ready",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "phases",
          "shared_agent_count",
          "phase4_agent_count",
          "components",
          "audit",
          "metrics"
        ],
        "title": "OperationalStatus",
        "type": "object"
      },
...
      "SecurityPolicyRead": {
        "additionalProperties": false,
        "properties": {
          "auth_rate_limit_requests": {
            "title": "Auth Rate Limit Requests",
            "type": "integer"
          },
          "idempotency_ttl_seconds": {
            "title": "Idempotency Ttl Seconds",
            "type": "integer"
          },
          "max_inflight_requests": {
            "title": "Max Inflight Requests",
            "type": "integer"
          },
          "max_request_bytes": {
            "title": "Max Request Bytes",
            "type": "integer"
          },
          "production_transport_security": {
            "title": "Production Transport Security",
            "type": "boolean"
          },
          "rate_limit_requests": {
            "title": "Rate Limit Requests",
            "type": "integer"
          },
          "rate_limit_window_seconds": {
            "title": "Rate Limit Window Seconds",
            "type": "integer"
          },
          "trust_proxy_headers": {
            "title": "Trust Proxy Headers",
            "type": "boolean"
          },
          "trusted_host_count": {
            "title": "Trusted Host Count",
            "type": "integer"
          }
        },
        "required": [
          "max_request_bytes",
          "rate_limit_requests",
          "auth_rate_limit_requests",
          "rate_limit_window_seconds",
          "max_inflight_requests",
          "idempotency_ttl_seconds",
          "trusted_host_count",
          "trust_proxy_headers",
          "production_transport_security"
        ],
        "title": "SecurityPolicyRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AstraPath Phase 1-5 API. Contracts are versioned under /api/v1 and support exactly two roles: student and admin.",
    "title": "AstraPath API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/operations/audit/verify": {
      "post": {
        "operationId": "verify_audit_api_v1_admin_operations_audit_verify_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerification"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify Audit",
        "tags": [
          "phase5-operations"
        ]
      }
    },
...
    "/api/v1/admin/operations/security-policy": {
      "get": {
        "operationId": "security_policy_api_v1_admin_operations_security_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityPolicyRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Security Policy",
        "tags": [
          "phase5-operations"
        ]
      }
    },
    "/api/v1/admin/operations/status": {
      "get": {
        "operationId": "operational_status_api_v1_admin_operations_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Operational Status",
        "tags": [
          "phase5-operations"
        ]
      }
    },
...
  }
}